requirements.txt
matplotlib==3.9.2
monai==1.3.2
nnunetv2==2.5.1
numpy==2.1.2
opencv_python==4.10.0.84
Pillow==10.4.0
pycocotools==2.0.8
scipy==1.14.1
SimpleITK==2.4.0
scikit-image==0.21.0
timm==1.0.9
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.2
wandb==0.18.7
click>=7.1,!=8.0.0
gitpython>=1.0.0,!=3.1.29
protobuf>=3.19.0,<6,!=4.21.0,!=5.28.0
psutil>=5.0.0